move spreadsheet code into a module
